+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.1] - 2019-04-30
+### Changed
+- Function from `is_mr_valid` to `check_if_mr_is_valid`.
+
+### Fixed
+- Exits with an error (1) if MR exists already, should exit without an error (0).
+
 ## [0.5.0] - 2019-04-30
 ### Added
 - Tag docker image with version from `setup.py`.